systemd-boot to go str8 to windows and more easily change params, terminal framebuffer
made it better font (but still can't show ❯) and using fbset to make it
fullscreen, not be limited to the size of the smaller mirrored screen.
and add some EFI tools for good measure. and r8169 to finally end the
enp4s0: rtl_rxtx_empty_cond == 0 woes? and vscode bad