PYTHON_HISTORY [new in 3.13](https://github.com/python/cpython/pull/13208)
thanks to
https://unix.stackexchange.com/questions/630642/change-location-of-python-history
- don't use a new folder cause 1) no need for just 1 file 2) python won't make
it for you, will just fail to write
history. [`xdg-ninja`](https://github.com/b3nj5m1n/xdg-ninja) is good, we need
more all-in-one places for doing config like this. but it's more or less just a
community thing

will need a place for these lines in zshrc /shrug

+ try some LSP servers
+ bat hard-wraps, less -S is nice